fix(ci): require the gate's classifier to be defined at top level
Split first, then the fix, per GitHub issue #1803.

The split mirrors the module split of two commits ago, with the seam that
one did not have: `gateFixture`, `installedBashes` and
`legacyFunctionSource` are used by both halves, so they move to
check-sentry-suites-in-ci-gate-fixtures.mjs rather than into either. The
extraction tests move to check-sentry-suites-in-ci-gate-extract.test.mjs.
Partitioned into top-level blocks with the partition asserted to reassemble
the original byte for byte before writing, and proven neutral by the sorted
case-name diff: 48 cases before, 48 after, identical names, on both
interpreters.

Then the fix. `trimStart()` tested indentation, which is not the same
question as top level, and Codex is right that the difference matters. A
classifier defined inside another function at column 0 passed that check,
so the probe extracted it and reported verdicts — including
`root-tooling-scripts` for an unallowlisted alias out of deliberately wrong
logic — for a function the gate never defines where it is called. The
routing test would have been green on a gate that fails at runtime.

Nothing in the toolchain prevented the layout. This repo has no shell
formatter: .trunk/trunk.yaml enables shellcheck, which is a linter, and
shellcheck reports no issue on a nested definition. An earlier claim in
this PR that "the formatter leaves it as written" rested on running
`trunk fmt` over an untracked file and reading the no-op as a decision;
that reasoning was wrong and the reply has been corrected.

The check asks bash about enclosure rather than measuring indentation:
everything before the definition's header parses as a complete script
exactly when that header is at top level, and fails to parse when it sits
inside a function, an `if`, or any other compound command. One `bash -n`
spawn, no execution of the gate, no brace counting — 13ms against the real
gate. The real classifier still reads as top level on both interpreters,
asserted in the same test.

Refs #1779
Refs #1803
