+++ scripts/check-sentry-suites-in-ci.test.mjs
@@ -97,7 +97,9 @@ import { countLines, HARD_CAP } from "./file-size-watchlist.mjs";
 // the line cap. Importing it registers its `test()`s in this run, so the single
 // `node scripts/check-sentry-suites-in-ci.test.mjs` CI step runs them too.
 import "./check-sentry-suites-in-ci-lifecycle.test.mjs";
-// The gate-routing probe's own invariants, split out for the same reason.
+// The gate-routing probe's own invariants, split out for the same reason: how
+// the classifier is lifted out, and how it is re-run once lifted.
+import "./check-sentry-suites-in-ci-gate-extract.test.mjs";
 import "./check-sentry-suites-in-ci-gate-probe.test.mjs";
 import {
   CI,
@@ -976,6 +978,8 @@ test("the checker's own files stay under the file-size hard cap", () => {
     "scripts/check-sentry-suites-in-ci-gate-probe.mjs",
     "scripts/check-sentry-suites-in-ci-gate-probe.test.mjs",
     "scripts/check-sentry-suites-in-ci-gate-extract.mjs",
+    "scripts/check-sentry-suites-in-ci-gate-extract.test.mjs",
+    "scripts/check-sentry-suites-in-ci-gate-fixtures.mjs",
     "scripts/check-sentry-suites-in-ci-lifecycle.test.mjs",
   ];
   const over = files
